chore(auth-providers): switch to vitest (mostly) (#9869)
This PR updates the auth providers with the exception of two:
1. `@redwoodjs/auth-firebase-web`
It contains jest specific code for specifically resolving uuid to CJS.
I'm sure it's possible to switch it over but I have not yet tried.
3. `@redwoodjs/auth-dbauth-web`
I encountered errors that WebAuthn was not supported within the test
environment. Even with the `jsdom` environment set.