chore(esm): convert `@redwoodjs/cli-helpers` to ESM (#9872)
This PR converts `@redwoodjs/cli-helpers` to a dual ESM/CJS package like
https://github.com/redwoodjs/redwood/pull/9870 did for
`@redwoodjs/project-config`. I didn't do anything differently so see
that PR for details.

As soon as I converted it, the Jest tests stopped working so
@Josh-Walker-GM and I bundled both changes into one PR here.

With this PR, we should be able to get all the tests in
https://github.com/redwoodjs/redwood/pull/9863 passing (about ~15 (out
of 1000+) are skipped right now) because Vitest will be able to mock
`fs` since it's imported instead of required. That'll be really nice
since converting the CLI to ESM will then be possible. I'll confirm that
after making this PR. Even if they don't this'll still have been
worthwhile.

---------

Co-authored-by: Josh GM Walker <56300765+Josh-Walker-GM@users.noreply.github.com>