chore(server): dedupe web serve logic (#9884)
I'm not attached to the names of anything or the package structure, etc.

First PR to clean up the logic around our servers. This one tackles the
web server since there was just less to do. The goal was to not have any
duplicated logic anywhere specific to any entry point so that we can fix
bugs in one place. Code was duplicated while we were experimenting but
those features are going to go out soon.

@dac09 I went with the adapter package structure we started on. I needed
to establish this for v7 but hoping it could also provide something we
could build on going forward so would love to know what you think.

Handing off to Tobbe for review and help.

- I need help with types. They're marked with `// @ts-expect-error TODO`
- We need to check that the index files I introduced aren't barrel
files/propagating a pattern we're trying to get rid of. In general, how
do we export stuff for testing but not for use?
- The `server.test.mts` tests still need more updating. I've started on
the important ones that ensure each entry point is configured correctly
but some snapshots are still specific to my system etc. I would
recommend not spending time on them yet
- Again just in general this isn't 100% done so I probably missed some
things; lemme know!