fix(cli): Prevent caching blank information about plugins (#9942)
**Problem**
Currently we cache command information which we can later use to show
help information without the need to fully load the plugin back into
memory. This was done for performance and also to avoid the need to have
plugins installed to show help output - otherwise it would not have been
possible to have `rw storybook` be a plugin and show it's help
information before installation.

This was highlighted during the development of the zenstack plugin for
the redwood CLI.

**Changes**
1. We no longer add cache entries where we gathered no useful data.
Previously we would have written an entry like `command: { aliases:
undefined, description: undefined }` which then would have prevented
showing any help information when the cache was used. Now no cache entry
will be present and the plugin will have to be loaded into memory to
show the correct help output.

---------

Co-authored-by: Tobbe Lundberg <tobbe@tlundberg.com>