feat(server): dedupe api server code, make host configurable (#9948)
Follow up to https://github.com/redwoodjs/redwood/pull/9884. Does
something similar for the api server but I didn't go quite as far.

The crux of this PR is: we duplicated code between
`@redwoodjs/api-server` and `@redwoodjs/fastify` while we were working
on experimental features. This PR brings the changes from
`@redwoodjs/fastify` into `@redwoodjs/api-server`.

There's still a few types I need help with, and there's some duplicated
helper-like function code that I have some ideas around deduplicating,
but here's a summary of the changes:

- Removes `ansi-colors`

We only use one colors package across the framework now, `chalk`. This
saves us ~27 kBs. I'm not attached to chalk or anything, but it's used
by far more of the framework and we don't need two colors packages.

- Adds the ability to configure the api/web host and port where possible
in many ways

In working on Docker and with "severful" deploy providers, many of them
expect to be able to configure the host and port the server is listening
at. Right now in Redwood you really can't configure the host at all.
I've added the ability to do this in the ways they'd all expect:

- Via CLI flag (`--apiHost`, or just `--host` depending on the entry
point)
  - Via env var (`REDWOOD_API_HOST`, etc)
  - Via TOML (`[api].host`)

And of course if none of these are set, there's our default which is
based on `NODE_ENV`.

To do this, I had to remove the default for host from
`@redwoodjs/project-config`. The api's value was never hooked up but
there may be some code that depended on the web one. I've tested dev and
CI tests dev though so I'm fairly certain that things still work out of
the box.

- Changes some file names

`@redwoodjs/api-server` had a file called `index.ts` that wasn't the
index (e.g. `"main"` in the package.json didn't point to it). This file
was really the entry point to the package's primary bin, `rw-server`, so
I named it `bin.ts`.

- `@redwoodjs/api-server` uses the package.json exports field

If there's one explicitly breaking change here, it's this. I know some
startups were using this package's default export in a hacky way to
create their own server file. I'll have to highlight this in the release
notes and emphasize the move to `createServer`.

This package also provides the logFormatter bin. You can still run the
bin, but I'm not sure if other packages are depending on parts of the
code in this directory in an indirect way (I couldn't find any explicit
imports).

- Most fastify plugins don't have to be awaited

`fastify-raw-body` is the only plugin that has to be awaited and we were
inconsistent with all the others (some packages awaited them; some
didn't). I've consistently not awaited those that don't need to be, but
I have moved them all above our logic which is important since fastify's
loading is graph-based.

- All code in `@redwoodjs/api-server` is a fastify plugin

Most of the plugins were in a kind of decorator `with` style where you'd
pass the object along a series of function calls that'd mutate it. This
isn't how fastify prefers to do things; it has a strong encapsulation
model built around plugins. Making our functionality plugins goes with
instead of against fastify's way of doing things, and we should stop
getting "plugin already registered" errors.

Todos after this PR is merged:

- [ ] Update studio which may be affected?

---------

Co-authored-by: Tobbe Lundberg <tobbe@tlundberg.com>