chore(api-server): switch to vitest (#9929)
This PR updates the `api-server` unit tests to use vitest.

This was the first time we have had to convert over a test which used
virtual modules. The vitest docs on their support for this can be found
here: https://vitest.dev/guide/mocking.html#virtual-modules. I went with
menu option 2 - having a vite plugin that resolves the virtual module
and then mocking that module inline in the test file. Happy to discuss
this and go with a different approach. I'm not overly attached to one
way of doing things over another. If people have good reasons to suggest
using a particular style then it would be great to hear.