fix(docker): compose dev and prod (#9982)
**Problem**
By default when once you have run the experimental setup command for
docker, the compose setup for dev and prod doesn't quite work.

**Changes**
1. We add missing `fs-extra` dependency for the `@redwoodjs/web-server`,
it is a runtime dependency.
2. We add a `REDWOOD_API_HOST=0.0.0.0` env var for the dev setup. This
is required for the vite dev server to proxy requests correctly.

**Notes**
@jtoar I'm not overly sure how to test the compose with the updated
dependency of the web-server package. Adding the dependency to the web
side itself did work so I would imagine this works fine too.