7.0 RC: Remove hardcoded check for `session.id` (#10013)
Fixes bug when User table had a primary key other than `id`. Shout out
to @will-ks for finding this!

### Impact 

For apps which had a primary key other than `id`, all users will be
logged out on their next request after this is deployed.

Not sure if we consider that breaking? But it is 7.0 so anything goes!

Closes #10005