feat(server): add docs on the server file (#10019)
This PR adds docs for the server file. I opted to add them to the Docker
doc since that's where the idea originated from and most serverful
deploys that aren't serverless are Docker-based (the only one that I can
think of that isn't is baremetal).

I also updated the app configuration and realtime docs. It seems like
the realtime setup command should update the graphql function, but that
was out of scope for this PR.