fix(realtime): update logic for including sseLink (#10025)
Since realtime went out of experimental, the logic for including the SSE
Link in the bundle needs to be updated. We can't just use TOML config
anymore as a tell since there's none. Right now all I can think of as a
real source of truth is if the user has the realtime package on the api
side.