chore(ci): add changelog check to ci (#9989)
This PR is a spike on improving the Changelog in our release process. It
adds a workflow to our CI that checks to see if a PR updated the
CHANGELOG.md file.

In pairing on release, the hardest part is the release notes. If release
is going to be more approachable and sustainable, we need to spread the
burden of writing release notes, and the best person to write the
release notes for a particular PR is the person who wrote it or the
person who reviewed it. And the best time to do it is before it's
merged! (Not weeks or months after.)

I like reading [esbuild's release
notes](https://github.com/evanw/esbuild/blob/main/CHANGELOG.md) and
tried to find out how Evan Wallace writes them. It seems like he just
adds a bullet point to the CHANGELOG.md file as the last commit before
merging a PR he writes or reviews. I'm not sure if it gets much simpler
than that, so I thought we could start there.

- If a user doesn't update the changelog:

Failing run:
https://github.com/redwoodjs/redwood/actions/runs/7850731400/job/21426621517?pr=9989

  Failing check:

<img width="675" alt="image"
src="https://github.com/redwoodjs/redwood/assets/32992335/97a1773f-317e-4c11-9ec1-2a73c3c40d96">

  Instructions in the check:

<img width="1532" alt="image"
src="https://github.com/redwoodjs/redwood/assets/32992335/21ec9ecf-263f-46dc-9632-00f1ff8cc084">

Some feedback I'm anticipating is: those instructions are hard to find,
this check should comment in the PR. Agree that if you don't know where
to look, you won't find them. But commenting in the PR complicates the
logic and get noisy fast (we've disabled every bot that comments in
threads). One more point: I think updating the Changelog should be on
the burden of maintainers, so as long as we know the process it should
work out. But happy for any feedback or ideas!