fix(server): prefix port/host with api, fix logging (#10035)
Last fix for v7 I noticed while trying to get
https://github.com/redwoodjs/redwood/pull/10034 done. In talking to Josh
about his experiences deploying to Fly etc and some of the pain points
around that, I may introduce a `createServer` for the web side in a
minor. If they're in the same file (which they will be for now), parsing
argv with just port and host will be a problem. So to avoid that I've
made it so that it's apiPort and apiHost.