fix: update package.json metadata
looks like i resolved these for the main branch when i merged the
release/major/v7.0.0 branch into main, but not for the release branch
and tag (and now next branch). the only real bug here is the
peerDependency version not being updated. that's probably because of an
error in the release script which i'll fix on main