fix(render): reduce memory and handle server file  (#10055)
Made some improvements to the render deploy command while debugging an
issue with realtime on their platform (see
https://community.redwoodjs.com/t/redwood-realtime-on-render-com/5796
for the original issue). Note that while this PR fixes the original
issue (subscriptions is not an allowed operation), realtime itself still
doesn't work as expected. Currently trying to isolate render if it is
specific to them. But these are valid fixes in the meantime:

- should improve the memory usage problem by 1) requiring
cli-data-migrate to be a devDependency prior to deploy and 2) not
importing from the cli's lib which brings in the whole world 3) not
using execa's shell option which isn't recommended
- render will now use the server file if it's there