Add support for additional env var files (#9961)
Fixes #9877

Adds a new middleware step to the CLI booting up, which looks for
`--include-env` and includes `.env.[file]` to the list of files to look
at.

Also generally adds a .env var based on the `NODE_ENV` - I could take
this or leave it, but I was in the space.

![image](https://github.com/redwoodjs/redwood/assets/49038/267c9f8d-0b4d-4c83-8607-712b837e3cfc)

---------

Co-authored-by: Tobbe Lundberg <tobbe@tlundberg.com>
Co-authored-by: Dominic Saadi <dominiceliassaadi@gmail.com>