chore(realtime): switch to vitest (#10146)
This PR switches out the `@redwoodjs/realtime` to use vitest.

There is a hack here to get it to work with the `@envelop/testing`
package we use as a testing utility provider. That package uses
`jest.fn()` and other jest functions although it does not depend on jest
(peer or otherwise). You can see in the `vitest.setup.mts` that I have
simply added `globalThis.jest = vi`. I am not thrilled with this, it
also has a type error I have marked as expected. I think we should
clarify with the Guild on this?