chore(ci): refactor out CLI smoke tests (#10206)
Continuation of the CI improvements started in
https://github.com/redwoodjs/redwood/pull/10205. Right now we run the
CLI smoke tests as a part of the Playwright smoke tests. This means
we're running the CLI smoke tests twice as much as we need to since the
Playwright smoke tests have a matrix that depends on the bundler (vite
or webpack) and it makes CI take longer as a whole since the CLI smoke
tests have to wait till the Playwright smoke tests finish to start
running. This PR refactors out the CLI smoke tests into their own
workflow.