RSC: Inline dependencies during rsc build (#10217)
See the comment in the code as:
Externalize every file apart from node built-ins. We want vite/rollup to
inline dependencies in the server bundle. This gets round runtime
importing of "server-only". We have to do all imports because we can't
rely on "server-only" being the name of the package. This is also
actually more efficient because less files. Although, at build time it's
likely way less efficient because we have to do so many files.

---------

Co-authored-by: Tobbe Lundberg <tobbe@tlundberg.com>