fix(rsc): Preinit client component css for server components in production (#10210)
**Problem**
Currently RSC during `yarn rw serve` would not load the CSS of nested
client components. In development, things appear to work fine without
additional manipulation due to the vite dev server.

**Approach**
During build, we insert
[preinit](https://react.dev/reference/react-dom/preinit) calls into
server components that cause the browser to load all CSS that is needed
for any client component that the server component will cause to be
loaded.

**Changes**
1. Update the RSC analyse plugin to construct a map between the
component id and the component import ids.
2. Add an additional plugin to the RSC server build step. This plugin
analyses the client build manifest to gather all the known CSS assets.
It then inserts `preinit` calls to load CSS assets into server
components that would require them because they load client components
that have associated CSS assets.

**Notes**
1. The tests added here are really only ensuring that we do not make a
change to the behaviour that we currently have. There's a bit of a limit
as to what we can do when we simply pass known values to a test rather
than run these in a more E2E style fashion in a fully fledged RSC test
project - with known src, dist, and browser rendering.