chore(tests): avoid needlessly matching on `src` (#10228)
When I updated projects to use vitest I kept some instances where we
used `run src` and then that got copied to nearly everywhere else. This
isn't all the useful as when you are debugging and passing a file path
as it runs all tests since the `src` matches all tests.