chore(rsc): upgrade to react canary (#10194)
Want to see how CI reacts to upgrading to React canary. ~I may have
over-upgraded (i.e. some fixtures and maybe the docs don't need to be
upgraded), but let's see what happens.~

This upgrade was complicated by
https://github.com/facebook/react/pull/27436 which added a poisoned
import to `react-server-dom-webpack/server`, meaning that if you tried
to import from that entry point without the "react-server" condition, it
would throw.

In a much-earlier PR, Sebastian mentions there generally being two ways
to do things:

> Either you can have the server code prebundled using Webpack (what
Next.js does in practice) or you can use an unbundled Node.js server
(what the reference implementation does).

See https://github.com/facebook/react/pull/26172.

This PR goes with the former, prebundling
`react-server-dom-webpack/server` with the "react-server" condition so
that we can avoid having to specify it at runtime. It's hard to tell if
this the better option long-term, but since we're keen on getting rid of
the worker right now, this moves us closer in that direction. The
alternative solution was to move decoding the reply into the worker
where the condition is specified.

---------

Co-authored-by: Tobbe Lundberg <tobbe@tlundberg.com>