fix(cli): Update build and serve to handle the absence of the api side (#10265)
**Problem**
This PR address the issue raised in #10264 which highlighted that `yarn
rw build` would fail for projects that have removed the `api` directory.

**Changes**
1. The `yarn rw build` command now checks for the existence of the
prisma schema file. It will not generate the prisma client if the file
does not exist and it will not trigger prerendering if the schema file
does not exist.
2. Updates the checks performed by `yarn rw serve` to handle the case
when there is not API side a little nicer. It will offer more specific
error messages when you run `yarn rw serve api` without an API side and
it will tell you to run `yarn rw serve web` instead of `yarn rw serve`
if you have no API side.

**Concerns**
1. I'm not sure about adding checks for "does the API side even exist"
anywhere we previously assumed it did. For now these checks only needed
to happen in 2 places so I'm not that bothered about them but I can see
how it could spiral out of control.