chore(telemetry): build with esbuild (#10360)
As @jtoar pointed out in #10323 we load and depend on additional
packages as a result of building with our babel build config. This PR
switches out the `@redwoodjs/telemetry` package to use esbuild over
babel for building itself.

The changes here are:
1. Remove babel config and dependencies
2. Build with esbuild instead, adding `tsx` and `typescript` dev
dependencies
3. Build types too. We didn't previously do that, impact on package size
vs usefulness?

The resulting file changes in the dist output are that the existing
files change and that each file now also has a `.d.ts` and `d.ts.map`
companion.

Although this does reduce the number of files/packages loaded when you
run `yarn rw dev --help` it doesn't come with a noticeable performance
improvement. Still nice not to load stuff that we don't actually need.