chore(e2e): use tarsync for syncing framework state to project (#10379)
**Problem**
Working on #10378 has made me think that the e2e script is not really
behaving as we would want it to when it comes to syncing the framework's
state/dependencies to the test project.

Please see my comments in that PR for more info but in short, we were
getting versions of dependencies that we had upgraded away from. This
was then causing problems when the e2e tests were run.

**Changes**
1. Use `tarsync` in place of `deps` and `copy`. 
2. We can no longer assume the `rimraf` is available and so we use `dlx`
in that case.

**Notes**
I made the tarsync command within the e2e script verbose with `-v`. I
don't really care if it's verbose or not so happy to change that if
anyone has an opinion on it.

Using `dlx` might also be contentious but it feels natural to me when we
want to ad-hoc use the package.