chore: make crwa e2e test work across branches (#10437)
On main we run CI on every pull request. On next, we run it on every
push to the next branch. This gives whoever's doing release confidence
that cherry picking was done correctly.

There's one CRWA test that doesn't translate across branches because
it's a snapshot of the package's version. See
https://github.com/redwoodjs/redwood/actions/runs/8607313332. This PR
updates that test so that it checks that the version string matches a
RegExp instead of a snapshot.