chore: remove aws-lambda
This removes `aws-lambda` as a `devDependency` from the packages it was
being referenced in.

The `aws-lambda` package is actually a CLI for deploying lambdas. The
`@types/aws-lambda` package provides type definitions for what AWS makes
available at run-time (i.e. the `"aws-lambda"` import path isn't meant
to exist on disk).