chore: remove aws-lambda (#10450)
This removes `aws-lambda` as a `devDependency` from the packages it was
being referenced in.

The `aws-lambda` package is actually a CLI for deploying lambdas. The
`@types/aws-lambda` package provides type definitions for what AWS makes
available at run-time (i.e. the `"aws-lambda"` import path isn't meant
to exist on disk).

I searched the repo for usages of the `lambda` bin it provides and it
doesn't look like there are any, so this was probably just added by
mistake long ago

---------

Co-authored-by: Josh GM Walker <56300765+Josh-Walker-GM@users.noreply.github.com>