fix(cli): avoid `npx` during upgrade command (#10479)
**Problem**
Fixes #10466.

During `yarn rw upgrade` we check the version of `npx` installed. It's
entirely possible that `npx` is not available or installed.

We only do this check because we have to handle dedupe differently
between yarn v1 and yarn >v1. We specify that redwood projects should be
using yarn v4 using `"packageManager": "yarn@4.1.1"` in the
`package.json`. Therefore when following the recommended setup users
should not be using yarn v1 in their redwood projects.

**Changes**
1. Avoid `npx` version check.
2. Skip dedupe step if for some reason have yarn v1. When this happens
we log a warning message to tell the user to run a command manually to
dedupe.